# services/docs-app/openapi/web-api.yaml
openapi: 3.0.3
info:
  title: RecipeBook API
  version: 0.7.0
  description: >
    # About

    Something about this API...
servers:
  - url: http://127.0.0.1:8080/api/v1/
    description: local deployment of web-api
components:
  securitySchemes:
    AccessTokenAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    CustomUnit:
      type: object
      properties:
        name:
          type: string
          example: medium
        amount:
          type: number
          example: 123.0
        unit:
          type: string
          example: g
        product_id:
          type: integer
          example: 3
...
    JournalGroup:
      type: object
      properties:
        ui_index:
          type: integer
          example: 2
        name:
          type: string
          example: lunch
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
          example: 2
        entry_date:
          type: string
          example: "2022-11-04"
        entry_time:
          type: string
          example: "13:29:03"
        product_id:
          type: integer
          example: 6
        amount:
          type: number
          example: 450.5
        unit:
          type: string
          example: g
        journal_group_ui_index:
          type: integer
          example: 2
    CreateJournalEntryPayload:
      type: object
      properties:
        entry_date:
          type: string
          example: "2022-11-04"
        entry_time:
          type: string
          example: "13:29:03"
        product_id:
          type: integer
          example: 6
        amount:
          type: number
          example: 450.5
        unit:
          type: string
          example: g
        journal_group_ui_index:
          type: integer
          example: 2
    DeleteJournalEntryPayload:
      type: object
      properties:
        id:
          type: integer
          example: 2
    JournalEntryDetailed:
      type: object
      properties:
        id:
          type: integer
          example: 2
        entry_date:
          type: string
          example: "2022-11-04"
        entry_time:
          type: string
          example: "13:29:03"
        product_id:
          type: integer
          example: 6
        product_name:
          type: string
          example: Hamburger
        product_density:
          type: number
          example: 0.9
        amount:
          type: number
          example: 450.5
        unit:
          type: string
          example: g
        journal_group_ui_index:
          type: integer
          example: 2
        nutrients:
          type: object
          additionalProperties:
            type: number
          example:
            protein: 0.9
            dietary_fiber: 1.2
            energy: 18.0
            carbohydrate: 3.9
            sugars: 2.6
            fat: 0.2
        custom_units:
          type: array
          items:
            $ref: "#/components/schemas/CustomUnit"
    # user_nutrient
    UserNutrient:
      type: object
      properties:
        nutrient_id:
          type: integer
          example: 2
        is_featured:
          type: boolean
          example: true
        daily_target_amount:
          type: number
          example: 12.3
        ui_index:
          type: integer
          example: 7
    UserNutrientDetailed:
      type: object
      properties:
        nutrient_id:
          type: integer
          example: 45
        is_featured:
          type: boolean
          example: true
        daily_target_amount:
          type: number
          example: 12.3
        ui_index:
          type: integer
          example: 7
        nutrient_name:
          type: string
          example: vitamin_b3
        nutrient_daily_value:
          type: number
          example: 16
        nutrient_unit:
          type: string
          example: mg
        nutrient_group:
          type: string
          example: vitamins
        nutrient_parent_name:
          type: string
          example: null
    DeleteUserNutrientPayload:
      type: object
      properties:
        id:
          type: integer
          example: 2
...
paths:
...
  /journal/groups:
    get:
      tags:
        - journal
      summary: Get all journal groups
      security: 
        - AccessTokenAuth: []
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalGroup"
              example:
                - ui_index: 1
                  name: breakfast
                - ui_index: 2
                  name: lunch
                - ui_index: 3
                  name: dinner
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  /journal/groups/update:
    post:
      tags:
        - journal
      summary: Update all journal groups
      description: This request will replace all existing journal groups for a user with the ones in this payload
      security: 
        - AccessTokenAuth: []
      requestBody:
        content:
          "application/json":
            schema:
              type: array
              items:
                $ref: "#/components/schemas/JournalGroup"
            example:
              - ui_index: 1
                name: breakfast
              - ui_index: 2
                name: lunch
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  /journal/entry:
    get:
      tags:
        - journal
      summary: Get all journal entries
      security: 
        - AccessTokenAuth: []
      parameters:
        - name: entry_date
          in: query
          required: true
          schema:
            type: string
            example: "2022-11-04"
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntryDetailed"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  /journal/entry/create:
    post:
      tags:
        - journal
      summary: Create a journal entry
      security: 
        - AccessTokenAuth: []
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/CreateJournalEntryPayload"
      responses:
        "201":
          description: Created
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/JournalEntryDetailed"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  /journal/entry/update:
    post:
      tags:
        - journal
      summary: Update a journal entry
      security: 
        - AccessTokenAuth: []
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/JournalEntry"
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  /journal/entry/delete:
    post:
      tags:
        - journal
      summary: Delete a journal entry
      security: 
        - AccessTokenAuth: []
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/DeleteJournalEntryPayload"
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  /journal/nutrients:
    get:
      tags:
        - journal
      summary: Get all user nutrients
      security: 
        - AccessTokenAuth: []
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserNutrientDetailed"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  /journal/nutrient/upsert:
    post:
      tags:
        - journal
      summary: Upsert a user nutrient
      security: 
        - AccessTokenAuth: []
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/UserNutrient"
      responses:
        "201":
          description: Created
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/UserNutrientDetailed"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
  /journal/nutrient/delete:
    post:
      tags:
        - journal
      summary: Delete a user nutrient
      security: 
        - AccessTokenAuth: []
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/DeleteUserNutrientPayload"
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found